simulate/gc: extract BaseGcDataWalker into public header
Move BaseGcDataWalker and PtrRange out of simulate_gc.cpp into a new
include/daScript/simulate/simulate_gc.h so the GC data walker base can be
reused. Method bodies move out-of-line; behavior unchanged.
